* commit 'b3022ae10c79b0bebbbf9c854ea7086c97bcb081':
  Add a Transform to SSA phase to the optimizing compiler.

+++ compiler/optimizing/code_generator.h
@@ -211,7 +211,7 @@ class LocationSummary : public ArenaObject {
       : inputs(instruction->GetBlock()->GetGraph()->GetArena(), instruction->InputCount()),
         temps(instruction->GetBlock()->GetGraph()->GetArena(), 0) {
     inputs.SetSize(instruction->InputCount());
-    for (int i = 0; i < instruction->InputCount(); i++) {
+    for (size_t i = 0; i < instruction->InputCount(); i++) {
       inputs.Put(i, Location());
     }
   }
